Project salvo com ajustes de Funcionalidades como a Exclusao do atributo redundante email e muito mais

+++ database/migrations/2023_10_21_204525_create_categoria_funcionarios_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('categoria');
             $table->string('grau');
-            $table->float('salariobase');
+            $table->decimal('salariobase',9,2);
             $table->timestamps();
         });
     }
